remove snarkjs
