+++ Makefile
@@ -9,7 +9,7 @@ LDFLAGSSTRING +=-X 'main.BuildTime=$(BUILDDATE)'
 LDFLAGS := -ldflags "$(LDFLAGSSTRING)"
 
 es-node: build
-	cp -r ethstorage/prover/snarkjs build/bin
+	cp -r ethstorage/prover/snark_lib build/bin
 	mkdir -p build/bin/snarkbuild
 
 build:
