+++ ethstorage/prover/snark_lib/blob_poseidon2_verification_key.json
